Instructor setting page frontend completed but API integeration is pending yet
